LLM/requirements_llm.txt
fastapi
uvicorn
streamlit
requests
boto3
langchain-aws
langchain-core
python-dotenv
anthropic
pandas